Converted Exchanges to use Mechanize.
Using Webrat for both code and tests wasn't practical. The Webrat
Mechanize adapter wasn't working in the latest version, but the older
version wasn't working for the tests and also needed to be
monkey-patched. Using Mechanize for the Exchanges lets us use the latest
Webrat version and get rid of monkey-patches.